+++ include/animations.h
@@ -23,6 +23,7 @@ public:
   void animateDead();
   void animateNotification(const char* title = "", const char* message = "");
   void animateCoding();
+  void animateCommitHistory();
   
   // Reset animation frames when switching emotions
   void resetAnimation(EmotionState emotion);
@@ -42,6 +43,7 @@ private:
   unsigned long lastDeadAnim;
   unsigned long lastNotificationAnim;
   unsigned long lastCodingAnim;
+  unsigned long lastCommitHistoryAnim;
   
   int sleepyFrame;
   int thinkFrame;
@@ -56,6 +58,7 @@ private:
   int deadFrame;
   int notificationFrame;
   int codingFrame;
+  int commitHistoryFrame;
 };
 
 extern AnimationManager animationManager;